refactor ♻️ 결제수단 모달 컴포넌트명 변경
